move setup functions to nodeserver lib, no further changes
